requirements.txt
django>=5.0,<6.0
psycopg2-binary
pillow
python-dotenv
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.0
django-import-export
openpyxl
